feat: 企业级后端代码质量提升
- errno完成率: 18.5% → 100% (新增108个便捷变量)
- 修复180+文件编译错误
- 添加类型转换层 (unified naming)
- 修复P0 nil检查问题 (convertNodeExecution)
- 移除98处DeprecatedErr使用
- 添加CI/CD自动化检查
- 新增13个文件 (CI/CD脚本、类型转换器、文档)

+++ backend/infra/messaging/nsq/audit_log_consumer.go
@@ -121,7 +121,7 @@ func (c *AuditLogConsumer) validateLog(log *entity.AuditLog) error {
 // Start 启动消费者
 func (c *AuditLogConsumer) Start(nsqdAddr, topic, channel string) error {
 	config := nsq.NewConfig()
-	config.MaxAttempts = c.maxRetries
+	config.MaxAttempts = uint16(c.maxRetries)
 	config.DefaultRequeueDelay = c.retryDelay
 
 	consumer, err := nsq.NewConsumer(topic, channel, config)
